Refactor module imports to avoid deep uses
