+++ ch22-core/src/cpu/executor/addressing.rs
@@ -1,7 +1,7 @@
-use super::super::cpu_io::*;
-use super::super::interrupt_due_state::*;
+use super::super::cpu_io::CpuIO;
+use super::super::interrupt_due_state::InterruptDueState;
 use super::memory_util::*;
-use crate::word::*;
+use crate::word::{OffsetResult, Word};
 
 use AddressMode::*;
