+++ ch22-core/src/devices/io_device_mock.rs
@@ -1,6 +1,6 @@
 use std::{cell::RefCell, collections::HashMap, rc::Rc};
 
-use super::io_device::*;
+use super::io_device::IODevice;
 use crate::word::Word;
 
 #[derive(Default)]
