+++ ch22-core/src/devices/static_device.rs
@@ -1,4 +1,4 @@
-use super::io_device::*;
+use super::io_device::IODevice;
 use crate::word::Word;
 
 pub struct StaticDevice {
