+++ ch22-core/src/video.rs
@@ -1,2 +1,5 @@
-pub mod field_data;
-pub mod video_memory_access;
+mod field_data;
+mod video_memory_access;
+
+pub use field_data::Field;
+pub use video_memory_access::{CRTCRangeType, VideoMemoryAccess};
